requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.4
django-ckeditor==6.1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-json-ld==0.0.4
django-livereload-server==0.3.2
django-pagedown==2.2.0
django-shorts==1.1
model-bakery==1.3.2
Pillow==8.2.0
pytz==2021.1
rules==3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
